The circular dependency between Dimension and Dataset is now deleted. See previous commit for more details.
All the variables inside the __new__ method of the Dimension are now made private and the rest of the code is modified accordingly.

The setter function for the name of the dimension is deleted for now to see if it breaks any tests. If it does, we will need to come up with a better of setting the name of the dimension